views , modal  and refactoring

+++ backend/src/controllers/notes.ts
@@ -2,11 +2,16 @@ import { RequestHandler } from "express";
 import NoteModel from "../models/note";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+  const authentecatedUserId = req.session.userId;
+
   try {
     // throw new Error("Something went wrong");
-    const notes = await NoteModel.find().exec();
+    assertIsDefined(authentecatedUserId);
+
+    const notes = await NoteModel.find({ userId: authentecatedUserId }).exec();
     res.status(200).json(notes); // send notes back to client
   } catch (error) {
     next(error);
@@ -15,7 +20,11 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 
 export const getNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const authentecatedUserId = req.session.userId;
+
   try {
+    assertIsDefined(authentecatedUserId);
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid note ID");
     }
@@ -25,6 +34,14 @@ export const getNote: RequestHandler = async (req, res, next) => {
     if (!note) {
       throw createHttpError(404, "Note not found!");
     }
+
+    if (!note.userId.equals(authentecatedUserId)) {
+      throw createHttpError(
+        401,
+        "Not authorized. You can not access this note"
+      );
+    }
+
     res.status(200).json(note);
   } catch (error) {
     next(error);
@@ -44,12 +61,17 @@ export const createNote: RequestHandler<
 > = async (req, res, next) => {
   const title = req.body.title;
   const text = req.body.text;
+  const authentecatedUserId = req.session.userId;
+
   try {
+    assertIsDefined(authentecatedUserId);
+
     if (!title) {
       throw createHttpError(400, "Title is required");
     }
 
     const newNote = await NoteModel.create({
+      userId: authentecatedUserId,
       title: title,
       text: text,
     });
@@ -77,7 +99,10 @@ export const updateNote: RequestHandler<
   const noteId = req.params.noteId;
   const newTitle = req.body.title;
   const newText = req.body.text;
+  const authentecatedUserId = req.session.userId;
+
   try {
+    assertIsDefined(authentecatedUserId);
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid note ID");
     }
@@ -89,6 +114,13 @@ export const updateNote: RequestHandler<
 
     if (!note) {
       throw createHttpError(404, "Note not found!");
+    }
+
+    if (!note.userId.equals(authentecatedUserId)) {
+      throw createHttpError(
+        401,
+        "Not authorized. You can not access this note"
+      );
     }
 
     note.title = newTitle;
@@ -103,7 +135,10 @@ export const updateNote: RequestHandler<
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const authentecatedUserId = req.session.userId;
   try {
+    assertIsDefined(authentecatedUserId);
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid note ID");
     }
@@ -114,6 +149,14 @@ export const deleteNote: RequestHandler = async (req, res, next) => {
       throw createHttpError(404, "Note not found!");
     }
 
+    if (!note.userId.equals(authentecatedUserId)) {
+      throw createHttpError(
+        401,
+        "Not authorized. You can not access this note"
+      );
+    }
+
+    // await note.remove();
     await NoteModel.deleteOne({ _id: noteId }).exec();
 
     res.sendStatus(204);
